qmake ABI detection fix for NetBSD and OpenBSD
The ABI dection works on inspecting the ELF header which should contain
the values for ELFOSABI, with the macro ELFOSABI_NETBSD equals 2 and
ELFOSABI_OPENBSD equals 12. However, on these systems the ELF binaries
are built using 0 so detection will fail and default to Linux, preventing
to use the correct qmake for Desktop Qt. Therefore, on these systems default 0
to use the host OS (NetBSD or OpenBSD), otherwise Linux as before.

Change-Id: I293389980860977ba6c2ad9903edd567f0a5b9f0
Reviewed-by: Jake Petroules <jake.petroules@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>