QMake: Warn if build directory and source directory are not on the same level
We used to warn about this before, but users contributed patches that got
merged that removed these warnings for qmake versions where this bug is
fixed (QTCREATORBUG-10024).

Oswald says this to the topic:
"it's possible that *some* related problems were fixed (i remember for a fact
that i tried), but the fundamental underlying reason is inherent in
fileFixify(). fixing it basically means rewriting half of qmake, which simply
isn't going to happen at this point."

Task-number: QTCREATORBUG-16430
Change-Id: I78535273adacd40895e5bf203a2ea17dcc81716d
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>