Revert "TextEditor: Do not delete both quotes on backspace"
We have a more intelligent placing of double quotes, that should
eliminate some of the incorrect/unwanted placed closing quotes.

Additionally we introduced an animation and highlight of automatically
inserted text. So it should be definitely clear that the closing quote
is going to be removed.

If there is still someone who gets annoyed by this reintroduced behavior
there is now a setting that lets you control the deletion of
automatically inserted text.

So it is safe to reintroduce the deletion of closing quotes on
backspace.

This reverts commit 4a0b2039c0ed5c1f34efaf8f53dcf04ad9f384df.

Change-Id: I92500ac218dc4c2e07a3b0ad01b06e6baa2e2c6c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>