Debugger: Set sensible default size for memory view
By default the view is too small when it opens, and it must be resized
every time in order to see its content...

In addition, do not move the view next to the cursor. It is typically on
the right side-bar, causing the window to be clipped on the main window
edge.

Change-Id: Ib6c6f4a478fe5d571fc50d21d239cee591d0d3bf
Reviewed-by: hjk <hjk@theqtcompany.com>