ProjectExplorer: Add option to stop applications from the same build directory
Creator can stop running applications when a build starts. This prevents isssues
on windows where the running application locks files in the build directory, which
in turn will make the build fail.

In addition to the current options to kill all applications or just those from the
current project, add a new option to kill all applications of the current project
that live in the same build directory.

Task-number: QTCREATORBUG-16470
Change-Id: Ib4a74e2452055eaf5f9e384555be5ef14de1cd40
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>