UnitTests: Clean up line prefixer test
In TDD development is a standard advised that you setup you values in the
block, run what you want to test in the second and assert in third. We
should try to stick it because it makes reading the test easier.

It is advised too put the setup code after the test if possible.

I used const QList<QByteArray> inputChunks = {"hello\n"} instead of
const QList<QByteArray> inputChunks {"hello\n"} because it is always using
the initializer_list constructor.

Change-Id: I1573566efc1327a2d8681e9ae7bd0ad6ee182789
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>