Timeline: Fix event compression for non-default heights
The "id" member of OpaqueColoredPoint2DWithSize is then used as the
number of events compressed into one. To discern it from an actual
id it's also negated. Thus, it needs to be updated when adding an
event. This is not a problem for default height items because their
top is always at 0, so that the addition of the 'y' members doesn't
have an effect.

Change-Id: Ibd7218954e36661d7f729465430de89821a41a7f
Task-number: QTCREATORBUG-16763
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>