QmlDesigner: Implementing backend editor
This patch implements a new editor for the connection view called
"Backend".
This new editor allows creating properties in the root item for
types registered by qmlRegister(Singleton)Type in the project.

This will make the type available in the qml file and allows to use the
type/object as backend.

The "backend object" can be just declared or locally instanciated.

For singletons we just add or remove the import. A singleton object cannot
be added twice, since the import does already exist when it is added once.

Change-Id: I2443e431c8564a88cde7ed02fa87f3ab83ca9877
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>