QmlDesigner: set QT_QUICK_CONTROLS_CONF
If we have a qtquickcontrols2.conf we set QT_QUICK_CONTROLS_CONF
for the qml2puppet.

For this we have to forward the project to PuppetCreator.
I forwarded the project instead of just the path to qtquickcontrols2.conf,
because we most likely add more project specific features in the future.

Change-Id: I0a90636ca6903e28499018c8d769bb07f7be5bb6
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>