don't detect readlink a-priori
it's pointless and causes an unnecessary delay.
what's worse, we wouldn't actually know whether we found a compatible
(gnu) readlink, so the fallback would never get activated. this would be
the case on macos (the problem wasn't noticed so far, because on macos
we build a bundle which doesn't need the wrapper in the first place).

Change-Id: Ibf4c370d5d9e25064e80faeaf12eb5fa1bc68c20
Reviewed-by: Eike Ziller <eike.ziller@qt.io>