QmlDesigner: Add more items to component combobox
It makes sense to also add items that are not in a default properties
to the component comobox.

These are items like styling delegates:

Slider {
    handle: Item {
    }
}

or the items of e.g. a Flipable:

Flipable {
    front: Item {
}

Those items are not shown in the navigator (TODO) and it makes
sense to edit them if they were an inline component.

Change-Id: Iab5b6e0238625b5a486901fbf617911d06a34581
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>