iOS: also process stderr to look for QML debug port
This is where the "Waiting for connection ..." message usually appears.

In general this mechanism of first replacing the port in the message
and then parsing the message to figure out that the iostool (not the
application) is actually listening is very confusing and backwards,
but fixing this is material for the master branch.

Change-Id: I73d3c5a34482403d275e6ea7a9ad59996121b02e
Task-number: QTCREATORBUG-17141
Reviewed-by: Christian Stenger <christian.stenger@qt.io>