Fix CDB detection
To make sure CDB is really installed on a Windows Kit, we need to assure the
'inc' subdirectory is present. Before this patch, if, for instance, the
Windows 8.1 kit is present (including a Debuggers subdirectory), but without
CDB, the first test under "32-bit" qmake will yield a false positive,
preventing the following 32-bit kits to be tested, then to fail under the
64-bit qmake test because the 'inc' directory is inexistent, finally
bringing about an empty CDB_PATH variable.

Change-Id: I1d7cf3350f4a63a556ef106faa5b3f0593dcf8c8
Reviewed-by: Andrew Knight <andrew.knight@intopalo.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: David Schulz <david.schulz@qt.io>