Preserve non-ASCII characters in MSVC environment
For the MSVC compiler we parse the environment variables from the
output of a 'set' call. So far we didn't enforce any encoding, which
means that the current code page is used. Then we tried to parse this
with QTextCodec::codecForLocale, which is not necessarily the same.

Avoid this by forcing UTF-8 as output (code page 65001),
and parse as UTF-8.

Task-number: QTCREATORBUG-17219
Change-Id: Ia1a6fa56e5c3e1599bf545795c655bf9376bae5c
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>