Fix readability of TODO entries on dark themes
The TODO plugin provides default values for color-coding backgrounds
based on the severity of each item. These colors are rather light, and
there is just one default with no theme awareness.

It is a bug to use a theme-aware text color on a fixed-color background
-- just switch to a dark theme such as flat-dark and observe the
unreadability.

This patch simply ensures that the default settings have reasonable
readability. These defaults are still not theme-aware, but at least they
are consistent.

Change-Id: Ibf35a241c70e0f1ea001b55f84fd2e1dd1a0a2ea
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>