Qmake: Make sure to mark new ProFileNodes as "parse in progress"
ProFileNodes start out as "no parse in progress" and need to be told
that there is a "parse in progress" when created as part of the project
parsing.

Otherwise they will not notify their associated run configuration that
they are done parsing, which in turn keeps the run buttons in the UI
inactive.

Change-Id: Ic37a9f1ccb6beecbf37f4fc54747122572e52c90
Reviewed-by: Eike Ziller <eike.ziller@qt.io>